plc_extractor/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyModbusTCP==0.2.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
schedule==1.2.0
loguru==0.7.2
pyModbusTCP